Fixes - (rename, typechange)
