+++ src/auth/certificates/__tests/CompletedProtoWallet.ts
