+++ src/auth/certificates/index.ts
@@ -1,4 +1,4 @@
 export { default as Certificate } from './Certificate.js'
 export * from './MasterCertificate.js'
 export * from './VerifiableCertificate.js'
-export * from './__tests/MockWallet.js'
+export * from './__tests/CompletedProtoWallet.js'
